% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_add_logo.R
\name{gg_add_logo}
\alias{gg_add_logo}
\title{Add a Logo to a ggplot}
\usage{
gg_add_logo(
  plot,
  logo_path = jefe_logo(),
  x = 0.02,
  y = -0.04,
  width = 0.25,
  height = 0.2,
  top_margin = 3,
  right_margin = 3,
  bottom_margin = 10,
  left_margin = 3
)
}
\arguments{
\item{plot}{A \code{ggplot2} object.}

\item{logo_path}{A file path to the logo image. Defaults to the internal
Jefe logo included with the package.}

\item{x}{Horizontal position of the logo (0–1 scale).}

\item{y}{Vertical position of the logo (0–1 scale).}

\item{width}{Width of the logo relative to the plot (0–1 scale).}

\item{height}{Height of the logo relative to the plot (0–1 scale).}

\item{top_margin}{Top margin (in mm) added to the plot.}

\item{right_margin}{Right margin (in mm) added to the plot.}

\item{bottom_margin}{Bottom margin (in mm) added to the plot.}

\item{left_margin}{Left margin (in mm) added to the plot.}
}
\value{
A \code{ggplot2} object with the logo applied.
}
\description{
Overlays a logo onto a \code{ggplot2} object using \code{{cowplot}}. By default,
uses the packaged Jefe logo stored internally in the package.
}
\examples{
\dontrun{
p <- ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) +
  ggplot2::geom_point()

gg_add_logo(p)
}

}
