% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_fonts_jefe.R
\name{gg_fonts_jefe}
\alias{gg_fonts_jefe}
\title{Load and Register Google Fonts for ggplot}
\usage{
gg_fonts_jefe(
  title_font = "Merriweather",
  subtitle_font = "Source Sans 3",
  body_font = "Roboto"
)
}
\arguments{
\item{title_font}{A character string specifying the font used for titles.}

\item{subtitle_font}{A character string specifying the font used for subtitles.}

\item{body_font}{A character string specifying the font used for body text.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effects.
}
\description{
Downloads and registers Google Fonts for use in ggplot2 visualizations.
This function uses \code{{sysfonts}} to load fonts and \code{{showtext}} to enable
consistent rendering across devices.
}
\examples{
\dontrun{
gg_fonts_jefe()
}

}
