% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jefe_logo.R
\name{jefe_logo}
\alias{jefe_logo}
\title{Path to the packaged Jefe logo}
\usage{
jefe_logo()
}
\value{
A character string giving the path to the packaged logo image.
}
\description{
Returns the file path to the Jefe logo image stored in the package.
}
\examples{
jefe_logo()

}
